# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tendo API
  version: 1.0.0
  description: Initial owner setup, local login sessions, membership-checked household reads, household-scoped subjects and items, item completion receipts, and operational health endpoints. Setup is an operator-authorized bootstrap, not public visitor registration. Sessions are opaque server-side records carried in an HttpOnly cookie.
servers:
  - url: /
paths:
...
  /api/v1/auth/oidc:
    get:
      operationId: getOidcStatus
      summary: Read configured OIDC status
      tags: [Auth]
      security: []
      parameters: [{ $ref: '#/components/parameters/RequestId' }]
      responses:
        '200':
          description: OIDC availability; displayName is present only when enabled.
          headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/OIDCStatus' }, examples: { disabled: { value: { enabled: false } }, enabled: { value: { enabled: true, displayName: OpenID Connect } } } } }
        '400': { description: Malformed trusted forwarded metadata., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '421': { description: Authority mismatch., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  /api/v1/auth/oidc/identity:
    get:
      operationId: getOidcIdentity
      summary: Read whether the current account has a linked OIDC identity
      tags: [Auth]
      security: [{ SessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/RequestId' }]
      responses:
        '200': { description: Link status., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/json: { schema: { $ref: '#/components/schemas/OIDCIdentityStatus' }, example: { linked: false } } } }
        '401': { description: No valid session; stale cookie is cleared., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' }, Set-Cookie: { $ref: '#/components/headers/ClearStaleSessionCookie' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '404': { description: OIDC is disabled., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' }, example: { type: 'about:blank', title: Not Found, status: 404, code: oidc_disabled } } } }
        '503': { description: Identity lookup unavailable., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
  /api/v1/auth/oidc/start:
    post:
      operationId: startOidc
      summary: Start OIDC login or account linking
      description: Requires canonical Origin. Link requires an active session and current local password. Flow cookie is short-lived, HttpOnly and SameSite=Lax.
      tags: [Auth]
      security: []
      parameters:
        - { $ref: '#/components/parameters/RequestId' }
        - { name: Origin, in: header, required: true, schema: { type: string, format: uri }, example: https://tendo.example }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/OIDCStartRequest' }, examples: { login: { value: { purpose: login } }, link: { value: { purpose: link, currentPassword: a-long-unique-passphrase } } } } }
      responses:
        '200': { description: Authorization URL and flow cookie., headers: { Set-Cookie: { $ref: '#/components/headers/SetOIDCFlowCookie' }, X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/json: { schema: { $ref: '#/components/schemas/OIDCStartResponse' }, example: { authorizationUrl: 'https://idp.example/authorize?client_id=client&state=opaque' } } } }
        '400': { description: Malformed request or forwarded metadata., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '401': { description: Link session missing or current password invalid., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' }, Set-Cookie: { $ref: '#/components/headers/ClearStaleSessionCookie' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' }, example: { type: 'about:blank', title: Unauthorized, status: 401, code: invalid_credentials } } } }
        '403': { description: Origin is missing or foreign., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '404': { description: OIDC disabled., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' }, example: { type: 'about:blank', title: Not Found, status: 404, code: oidc_disabled } } }}
        '409': { description: Login requested while already signed in., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' }, example: { type: 'about:blank', title: Conflict, status: 409, code: already_signed_in } } }}
        '413': { description: Body exceeds 4 KiB., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }}
        '415': { description: Unsupported media type., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }}
        '421': { description: Authority mismatch., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }}
        '422': { description: Invalid purpose/body combination., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/OIDCValidationProblem' }, example: { type: 'about:blank', title: Validation Failed, status: 422, code: invalid_value, field: purpose } } }}
        '429': { description: Rate limit exceeded., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' }, Retry-After: { schema: { type: integer, minimum: 1 } } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }}
        '503': { description: Provider or persistence unavailable., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' }, example: { type: 'about:blank', title: Service Unavailable, status: 503, code: oidc_unavailable } } }}
  /api/v1/auth/oidc/callback:
    get:
      operationId: completeOidcCallback
      summary: Complete OIDC authorization code callback
      description: Always redirects to a fixed SPA destination with an empty body. Query values are never reflected or logged.
      tags: [Auth]
      security: []
      parameters:
        - { $ref: '#/components/parameters/RequestId' }
        - { name: code, in: query, required: false, schema: { type: string, maxLength: 4096 } }
        - { name: state, in: query, required: false, schema: { type: string, maxLength: 2048 } }
        - { name: error, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: error_description, in: query, required: false, schema: { type: string, maxLength: 4096 } }
        - { name: iss, in: query, required: false, schema: { type: string, maxLength: 2048 } }
      responses:
        '303': { description: Redirect to fixed SPA route; body is always empty., headers: { Location: { schema: { type: string, pattern: '^/(|login#oidcError=(cancelled|invalid_flow|identity_not_linked|identity_conflict|session_changed|authentication_failed|unavailable)|account(#oidc=connected|#oidcError=(cancelled|invalid_flow|identity_conflict|session_changed|authentication_failed|unavailable)))$' } }, Set-Cookie: { description: Flow cookie cleared; a session cookie is additionally set only on success. Multiple Set-Cookie header fields are possible., schema: { type: array, items: { type: string } } }, X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' }, Referrer-Policy: { schema: { type: string, const: no-referrer } } }}
        '400': { description: Malformed forwarded metadata., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
        '421': { description: Authority mismatch., headers: { X-Request-ID: { $ref: '#/components/headers/RequestId' }, Cache-Control: { $ref: '#/components/headers/NoStore' } }, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: tendo_session
      description: Opaque session cookie, 43 base64url characters. The cookie is named __Host-tendo_session and marked Secure when TENDO_PUBLIC_URL is HTTPS, and tendo_session without Secure when it is plain HTTP. Always HttpOnly, SameSite=Lax, Path=/, no Domain, 30-day absolute lifetime.
    SetupToken:
      type: apiKey
      in: header
      name: X-Tendo-Setup-Token
      description: Standard base64 encoding of 32 random bytes (44 characters ending in =). Operator generates with openssl rand -base64 32.
  headers:
    SetOIDCFlowCookie:
      required: true
      description: Ten-minute browser-binding cookie. HttpOnly; Path=/; SameSite=Lax; no Domain. HTTPS uses __Host-tendo_oidc with Secure; HTTP uses tendo_oidc.
      schema: { type: string }
      example: __Host-tendo_oidc=AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA; Path=/; Max-Age=600; HttpOnly; Secure; SameSite=Lax
    ClearStaleSessionCookie:
      description: Sent only when the request carried one session cookie naming an unknown, expired, or revoked session. Clears that cookie with an empty value and Max-Age=0, using the same name and attributes as SetSessionCookie.
      required: false
      schema: { type: string }
      example: __Host-tendo_session=; Path=/; Max-Age=0; HttpOnly; Secure; SameSite=Lax
...
    RequestId:
      description: Correlation ID echoed from a valid request header or generated by Tendo.
      schema: { type: string, pattern: '^[A-Za-z0-9._-]{1,64}$' }
      example: smoke-test_01
    NoStore:
      description: Prevents responses from being cached.
      schema: { type: string, const: no-store }
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller-supplied correlation ID.
      schema: { type: string, pattern: '^[A-Za-z0-9._-]{1,64}$' }
      example: smoke-test_01
...
  schemas:
    OIDCStatus:
      type: object
      additionalProperties: false
      required: [enabled]
      description: displayName is present exactly when enabled is true.
      properties:
        enabled: { type: boolean }
        displayName: { type: string }
    OIDCIdentityStatus:
      type: object
      additionalProperties: false
      required: [linked]
      properties: { linked: { type: boolean } }
    OIDCStartRequest:
      type: object
      additionalProperties: false
      required: [purpose]
      description: currentPassword is required when purpose is link and must be absent when purpose is login; otherwise the server returns 422 OIDCValidationProblem.
      properties:
        purpose: { type: string, enum: [login, link], x-enum-varnames: [OIDCPurposeLogin, OIDCPurposeLink] }
        currentPassword: { type: string, maxLength: 512 }
    OIDCStartResponse:
      type: object
      additionalProperties: false
      required: [authorizationUrl]
      properties: { authorizationUrl: { type: string, format: uri } }
    OIDCValidationProblem:
      type: object
      required: [type, title, status, code, field]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer, const: 422 }
        code: { type: string, enum: [invalid_value] }
        field: { type: string, enum: [purpose, currentPassword] }
      additionalProperties: false
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
      additionalProperties: true
...